use less to create button CSS
An example of the power of LESS for this kind of work. Previously
the buttons were pages of output CSS, cut and pasted and tweaked. With
less you can mostly just declare what you want things to look like and
it does the rest.

It's not perfect, the admin buttons need work, but you get the idea.