remove expendable.less, shift content other places
This file was only necessary because IE8/9 was getting sent
CSS files to long for them to handle. That issue was fixed in
moodle core, so moving these less vital styles back into their
logic places in table.less and button.less.

Also just deleted the rtl table stuff in favour of seeing what
bootstrap core come up with in v3.2 and the progress bar stuff
since it's crazy to include so much CSS for something that even most
admins will never see.

(I've checked to see if it's fixable in renderers, it's not, it's in
lib/weblib and is generally horrible. I did notice at the same time
that the progress bar on the new assignment edit screen hardcodes
Bootstrap 2.3 classes, so that's something to look out for).