various admin table cleanups
setting things to left-align is mostly a holdover from when text-align:
center was used to center things.

The rule .path-admin #assignrole .admintable is broken, it should have
missed it.

.roleassign micro-management not needed (though removing the
generaltable class from this layout table would help)

added a couple of bug numbers for things that can be removed
once the bug is fixed (one is already going into integration).

Made the current theme highlight in the chooser look nicer.

Removed some RTL versions of stuff that is currently commented out.

Cache table stuff was already 100% wide. The padding above the stuff
below the table was only needed because the tables own margin
was being removed in another rule intended to center the table.