Some workarounds that aren't necessary anymore
I *think* the way that btn and btn-default are done in Bootstrap3
shouldn't visually change the c1.btn class on the login block as
much, but for whatever reason I can't test it as I can't seem
to get the login block to appear on my test site.

I don't understand the tooltip one, but that CSS seems to already
be provided by Bootstrap.

The menu fix only seems to be needed for sites with a silly amount
of languages, and if I recall broke things for other menus anyway.