mapping BS3 forms to Moodle's HTML
This was one of the trickiest bits of getting Bootstrap 2.3 to
work with Moodle, and Bootstrap 3 totally changed the way this was
done. So I worked on the principle that if it was better to start from
scratch. So I commented out the entire forms.less file and slowly
started working my way through the file from top to bottom either
uncommenting bits or deleting them.

The bits that are still commented out at the bottom mean I've not
got around to them yet. Maybe they can be deleted, maybe they need
rewritten. There's almost certainly lots of niggly bugs left because
every form in Moodle is different for no reason but the basics of
responsive forms seem done so I thought I'd commit this.