use bootstrap classes on environment table
The table you see on updates. The changes are tiny but the render
is absolutely massive and scary with display and logic firmly
intertwined. I'm going to try to get the classes added to core
but in the meantime here's a find and replace version.

Getting it into core is complicated by the fact that text-error
and label-important got harmonized to -danger in Bootstrap 3.

Also removed the table-condensed rule to make the links more
finger friendly on mobile.