remove commented out admin form styles
These should all be covered by the recent changes in forms.less
which is the appropriate place for them anyway.

Also deleted the border on the device detect regex table since
it was right next to them and I just saw that my change to fix the
underlying problem was just integrated in 2.6 and master.