remove hardcoded color for login page details
This hardcode's #111 for a couple of things on the login page
which messes with bootswatches. I was going to move it to
the custom bootswatch but I'm not sure it's correct. It seems to be
making a link not look like a link, but I'm not sure why.

And the label tex should already inheret the correct color.