move rtl only CSS to their own files
Also added a switching mechanism based on the current language,
note that this will then cache RTL or LTR until the next theme
change so it's not suitable for sites using both directions.

There's still an issue with the columns on the front page, I
think there's some Moodle mechanism for RTL that's fiddling with
the block order and interfering with the basic Bootstrap grid.

edit: actually I think it was just the block swapping stuff at
the bottom of config.php that was causing that, looks better now
that I've deleted it anyway.