add a renderer for filemanager forms
Based on the upstream work that adds Bootstrap 2.3 classes
to the renderers.

There's a lot of code here for fairly small changes, but the way
this renderer is set up is kinda weird so I'm not sure it's
possible to do better. There's a couple of public methods that
use meta-programming to grab all the private methods on the class
that match a naming convention and process them. Possibly it can
be done more cleanly but I'd need to think about it a fair bit
more than I should have to.

Also, as a consequence of the way the templates are rewritten by
javascript I can't directly add the .form-control class so that's
being done via less.