add customcss setting
The search and replace stuff could be simplified I think.
When you set up a setting you provide Moodle with a default setting,
yet most of the themes I see seem to check if its set and then
manually add this default value. Perhaps you can ask Moodle for it,
since it already knows what it is.