simple settings class
Put the stuff for simplifying theme settings in a class to keep it
tidy, particularly as Moodle suggests functions should all start
with the full themename.

Use __DIR__ for locating theme local files in requires.

Also, dropped the over fancy and frankly unnecessary use of an
anonymous function and array_map in the CSS setting swap code for
a simple foreach, as it is simpler, shorter and more readable.