Style fatal errors better
To test try editing a URL to visit a section of the settings
that doesn't exist or a really high course id number. With
debug up full you see the debug info too.

Turns fatal errors into alert-dangers, and debug info into
pre.alert-infos.

The debug info just picks up on the (somewhat random) use of
notifytiny as a notification class. It also restyles it slightly
by rewriting the HTML, rather than add extra CSS just for debug
styles.

The fatal error stuff comes from deep within a scary renderer,
but by catching it in the box renderer I can divert it. This is an
interesting technique with possibiities for catching other stuff
that's either not in a renderer or in a really ugly renderer that
we'd rather not touch.