remove some hardcoded colors
Still a bunch left, see #206 for more info.

For some I already knew that the whole CSS wasn't used, so I
just deleted it completely.

Also removed the yui_fixes.less file which was left over from
before and not currently used, and the bootswatch/global which
I brought in by mistake because I thought it was a theme called
global, but it was just some setup files.