tighten CSS on form item min-width
We set a min-width on form fields so they line up neatly,
but the rule was wide enough to hit some other, more specialised
form items e.g. the pair of text fields within a table used to
specify device detection regexes in the the theme settings

Might need to keep an eye out for places this needs to be widened
again, but a quick check shows it mostly working.

closes #238